Made it so that the calculator handles float values properly

+++ calculatorC.cpp
@@ -3,25 +3,25 @@
 #include <stdlib.h>
 #include <string>
 void calculatorFunction();
-void calculatorFunction(int calculationResult);
+void  calculatorFunction(float calculationResult);
 
     calculatorC::calculatorC(){
         std::cout << "Calculate at your will! \n";
     };
 
-    int calculatorC::addNumbers(int firstNum, int secondNum){
+    int calculatorC::addNumbers(float firstNum, float secondNum){
             result = firstNum + secondNum;
             return result; 
         };
-    int calculatorC::subtractNumbers(int firstNum, int secondNum){
+    int calculatorC::subtractNumbers(float firstNum, float secondNum){
             result = firstNum - secondNum;
             return result;
         };
-    int calculatorC::multiplyNumbers(int firstNum, int secondNum){
+    int calculatorC::multiplyNumbers(float firstNum, float secondNum){
             result = firstNum * secondNum;
             return result;
         };
-    int calculatorC::divideNumbers(int firstNum, int secondNum){
+    int calculatorC::divideNumbers(float firstNum, float secondNum){
             result = firstNum / secondNum;
             return result;
         };
